- nasty float was causing calculation problems that would remove a pixel from the image's width / height
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@19842 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/etk/src/lib/etk_image.c
@@ -482,12 +482,12 @@ static void _etk_image_size_allocate(Etk_Widget *widget, Etk_Geometry geometry)
    {
       if (image->keep_aspect)
       {
-         float aspect_ratio;
+         double aspect_ratio;
          int image_w, image_h;
          int new_size;
 
          evas_object_image_size_get(image->image_object, &image_w, &image_h);
-         aspect_ratio = (float)image_w / image_h;
+         aspect_ratio = (double)image_w / image_h;
 
          if (geometry.h * aspect_ratio > geometry.w)
          {
